moved container_operators module
