+++ dags/grobid_build_image.py
@@ -1,10 +1,12 @@
 from airflow.models import DAG
 
+from sciencebeam_airflow.utils.airflow import add_dag_macro
+from sciencebeam_airflow.utils.container_operators import ContainerRunOperator
+
 from sciencebeam_dag_ids import ScienceBeamDagIds
 
 from sciencebeam_dag_utils import (
     get_default_args,
-    add_dag_macro,
     create_trigger_next_task_dag_operator,
     get_gcp_project_id
 )
@@ -13,8 +15,6 @@ from grobid_train_utils import (
     get_grobid_trainer_tools_image,
     create_grobid_train_validate_config_operation
 )
-
-from container_operators import ContainerRunOperator
 
 
 DEFAULT_ARGS = get_default_args()
