+++ dags/grobid_train_prepare.py
@@ -1,11 +1,13 @@
 from airflow.operators.bash import BashOperator
 from airflow.models import DAG
 
+from sciencebeam_airflow.utils.airflow import add_dag_macro
+from sciencebeam_airflow.utils.container_operators import ContainerRunOperator
+
 from sciencebeam_dag_ids import ScienceBeamDagIds
 
 from sciencebeam_dag_utils import (
     get_default_args,
-    add_dag_macro,
     create_trigger_next_task_dag_operator
 )
 
@@ -15,8 +17,6 @@ from grobid_train_utils import (
     create_grobid_train_validate_config_operation,
     DEFAULT_GROBID_TRAIN_FIELDS
 )
-
-from container_operators import ContainerRunOperator
 
 
 DEFAULT_ARGS = get_default_args()
