+++ dags/sciencebeam_dag_utils.py
@@ -3,7 +3,6 @@ import logging
 import os
 from datetime import timedelta
 from functools import partial
-from inspect import getmembers
 from pprint import pformat
 from urllib.parse import urlparse
 from pathlib import Path
@@ -342,22 +341,6 @@ def get_filtered_tasks(conf: dict, tasks: List[str]) -> List[str]:
     ]
     assert tasks[0] in set(filtered_tasks)
     return filtered_tasks
-
-
-def add_dag_macro(dag: DAG, macro_name: str, macro_fn: callable):
-    if not dag.user_defined_macros:
-        dag.user_defined_macros = {}
-    dag.user_defined_macros[macro_name] = macro_fn
-
-
-def add_dag_macros(dag: DAG, macros: object):
-    LOGGER.debug('adding macros: %s', macros)
-    for name, value in getmembers(macros):
-        if callable(value):
-            LOGGER.debug('adding macro: %s', name)
-            add_dag_macro(dag, name, value)
-        else:
-            LOGGER.debug('not adding macro, not callable: %s -> %s', name, value)
 
 
 def get_gcp_project_id():
